lift functions from postgres impl to catalog level
