+++ crates/iceberg-catalog/src/implementations/postgres/catalog.rs
@@ -14,19 +14,15 @@ use super::{
     },
     CatalogState, PostgresTransaction,
 };
-use crate::api::iceberg::types::PageToken;
-use crate::api::management::v1::{DeletedTabularResponse, ListDeletedTabularsResponse};
-use crate::implementations::postgres::dbutils::DBErrorHandler;
+use crate::implementations::postgres::tabular::list_tabulars;
 use crate::implementations::postgres::tabular::view::{
     create_view, drop_view, list_views, load_view, rename_view, view_ident_to_id,
 };
-use crate::implementations::postgres::tabular::{list_tabulars, DeleteKind};
-use crate::service::task_queue::tabular_expiration_queue::{ExpirationInput, TableExpirationTask};
-use crate::service::task_queue::TaskQueue;
+use crate::service::tabular_idents::{TabularIdentOwned, TabularIdentUuid};
 use crate::service::{
-    CreateNamespaceRequest, CreateNamespaceResponse, CreateTableRequest, DropFlags,
-    GetWarehouseResponse, ListFlags, ListNamespacesQuery, ListNamespacesResponse, NamespaceIdent,
-    Result, TableIdent, WarehouseStatus,
+    CreateNamespaceRequest, CreateNamespaceResponse, CreateTableRequest, DeletionDetails,
+    DropFlags, GetWarehouseResponse, ListFlags, ListNamespacesQuery, ListNamespacesResponse,
+    NamespaceIdent, Result, TableIdent, WarehouseStatus,
 };
 use crate::{
     api::iceberg::v1::{PaginatedTabulars, PaginationQuery},
@@ -41,9 +37,7 @@ use crate::{
     SecretIdent,
 };
 use iceberg::spec::ViewMetadata;
-use iceberg_ext::catalog::rest::ErrorModel;
 use std::collections::{HashMap, HashSet};
-use std::sync::Arc;
 
 #[async_trait::async_trait]
 impl Catalog for super::Catalog {
@@ -379,131 +373,21 @@ impl Catalog for super::Catalog {
         drop_view(table_id, drop_flags, transaction).await
     }
 
-    async fn list_soft_deleted_tabulars(
+    async fn list_tabulars(
         warehouse_id: WarehouseIdent,
-        catalog_state: Self::State,
+        list_flags: ListFlags,
+        catalog_state: CatalogState,
         pagination_query: PaginationQuery,
-    ) -> Result<ListDeletedTabularsResponse> {
-        let PaginatedTabulars {
-            tabulars,
-            next_page_token,
-        } = list_tabulars(
+    ) -> Result<PaginatedTabulars<TabularIdentUuid, (TabularIdentOwned, Option<DeletionDetails>)>>
+    {
+        list_tabulars(
             warehouse_id,
             None,
-            ListFlags::only_deleted(),
+            list_flags,
             &catalog_state.read_pool(),
             None,
             pagination_query,
         )
-        .await?;
-        Ok(ListDeletedTabularsResponse {
-            tabulars: tabulars
-                .into_iter()
-                .map(|(k, (ident, delete_opts))| {
-                    let i = ident.into_inner();
-                    let deleted = delete_opts.ok_or(ErrorModel::internal(
-                        "Expected delete options to be Some, but found None",
-                        "InternalDBError",
-                        None,
-                    ))?;
-
-                    Ok(DeletedTabularResponse {
-                        id: *k,
-                        name: i.name,
-                        namespace: i.namespace.inner(),
-                        typ: k.into(),
-                        warehouse_id: *warehouse_id,
-                        created_at: deleted.created_at,
-                        deleted_at: deleted.deleted_at,
-                        deleted_kind: deleted.deletion_kind.into(),
-                    })
-                })
-                .collect::<Result<Vec<_>>>()?,
-            next_page_token,
-        })
+        .await
     }
-
-    async fn expire_soft_deleted_tabulars(
-        warehouse_ident: WarehouseIdent,
-        catalog_state: Self::State,
-        expiration_q: Arc<
-            dyn TaskQueue<Task = TableExpirationTask, Input = ExpirationInput>
-                + Send
-                + Sync
-                + 'static,
-        >,
-    ) -> Result<()> {
-        let mut pagination = PaginationQuery {
-            page_size: Some(100),
-            page_token: PageToken::NotSpecified,
-        };
-        loop {
-            let mut transaction =
-                catalog_state.write_pool().begin().await.map_err(|e| {
-                    e.into_error_model("acquiring database transaction failed.".into())
-                })?;
-            tracing::debug!("Fetching new page");
-            let tabular_page = list_tabulars(
-                warehouse_ident,
-                None,
-                ListFlags::only_deleted(),
-                &mut *transaction,
-                None,
-                pagination.clone(),
-            )
-            .await?;
-
-            let next_page_token = tabular_page.next_page_token;
-            let tabulars = tabular_page.tabulars;
-
-            tracing::debug!(
-                "Fetched new page with '{}' items, starting to drop",
-                tabulars.len()
-            );
-
-            for (id, (_, delete_opts)) in tabulars {
-                let deleted = delete_opts.ok_or(ErrorModel::internal(
-                    "Expected delete options to be Some, but found None",
-                    "InternalDBError",
-                    None,
-                ))?;
-                // TODO: bail out or skip errors?
-                if deleted.deletion_kind == DeleteKind::Default {
-                    tracing::info!("Nor purge requested");
-                    drop_table(
-                        (*id).into(),
-                        DropFlags::default().hard_delete(),
-                        &mut transaction,
-                    )
-                    .await?;
-                } else if deleted.deletion_kind == DeleteKind::Purge {
-                    expiration_q
-                        .enqueue(ExpirationInput {
-                            tabular_id: *id,
-                            warehouse_ident,
-                            tabular_type: id.into(),
-                        })
-                        .await?;
-                }
-            }
-            transaction.commit().await.map_err(|e| {
-                tracing::warn!(?e, "failed to commit");
-                e.into_error_model("failed to commit transaction".into())
-            })?;
-            if let Some(next_page_token) = next_page_token {
-                pagination.page_token = PageToken::new_present(next_page_token);
-            } else {
-                break;
-            }
-        }
-
-        Ok(())
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-pub(crate) struct DeletionDetails {
-    pub(crate) deletion_kind: DeleteKind,
-    pub(crate) deleted_at: chrono::DateTime<chrono::Utc>,
-    pub(crate) created_at: chrono::DateTime<chrono::Utc>,
 }
