+++ crates/iceberg-catalog/src/service/mod.rs
@@ -12,7 +12,7 @@ pub mod token_verification;
 
 pub use catalog::{
     Catalog, CommitTableResponse, CreateNamespaceRequest, CreateNamespaceResponse,
-    CreateTableRequest, CreateTableResponse, DropFlags, GetNamespaceResponse,
+    CreateTableRequest, CreateTableResponse, DeletionDetails, DropFlags, GetNamespaceResponse,
     GetStorageConfigResponse, GetTableMetadataResponse, GetWarehouseResponse, ListFlags,
     ListNamespacesQuery, ListNamespacesResponse, LoadTableResponse, NamespaceIdent, Result,
     TableCommit, TableIdent, Transaction, UpdateNamespacePropertiesRequest,
