+++ crates/iceberg-catalog/src/service/tabular_idents.rs
@@ -63,7 +63,7 @@ impl<'a> TabularIdentBorrowed<'a> {
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub(crate) enum TabularIdentOwned {
+pub enum TabularIdentOwned {
     // TODO: TableIdent is from iceberg-rust, AFAIK, TableIdent and ViewIdent are the same, should we
     //       duplicate the type or use the same type and just accept it's called TableIdent?
     Table(TableIdent),
